requirements.txt
eyecite==2.6.9
reporters-db==3.2.58
PyMuPDF==1.23.0
pdfplumber==0.10.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy
pydantic
python-dotenv
structlog
requests
